Practice: Part 4b - beforeEach() w promises array
Create array of promises resulting from db .save() operation.

Use Promise.all(<array of promises>) to await a single Promise since
this method will return a single promise that resolves to an array of
the results of the input promises.

Do this since async callback in .forEach() loop actually is not
'await'ed in each iteration under the hood of forEach().