✨ Part 5c: Ex. 5.16 Unit test <AddBlogForm />
Check that the form calls the event handler it received as props with
the right details when a new blog is created.

NOTE 1: put `expect(<async function that updates component state>)` inside
testing-library/dom's async util `await waitFor(() => ...)` which itself
wraps react's `act()` to ensure that async call is awaited for settling
component states before testing and thus resolving `Warning: An update
to AddBlogForm inside a test was not wrapped in act(...).`

NOTE 2: To have async act() capability and get rid of error on
`waitFor()`: `TypeError: MutationObserver is not a constructor.`
upgrade package.json dependency specifications (including major
versions) with npm-check-update (ncu) utility.