✨ Part 7b: Ex. 7.8 implement useResource() hook
[App]
The `useResource` custom hook returns an array of two items just like
the state hooks. The first item of the array contains all of the
individual resources (an array) and the second item of the array is an
(service) object that can be used for manipulating the resource
collection, like creating new ones.

It uses useEffect hook to initialize resources (instead of exposing
a `getAll` service).